Change serializer to use email as username when creating User
Currently drfpasswordless doesn't set a username when
registering a User. This causes Django to throw an
exception when you try to register a second user
with no username, since username is a unique field.

This fix tries to retrieve the user by email first,
and creates one if not.

Technically should work for mobile numbers too.