Lock the password only, if the user was created
drfpasswordless locks the password each time, if the Token object
was created and not updated. This means that the user that uses the
Token login in the first time has no longer a password, because
it will deleted.

This commit changes the logic to only lock the password, if the user
was generated by drfpasswordless.